Move `cast`, `call`, and `handle` under `Process`
